+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - is_pos_int:  Checks to see if number is an integer and positive.
 
 ### Changed
+- mysql_stat_run:  Replaced datetime call with calls to gen_libs.get_date and gen_libs.get_time.
 - config/mysql_cfg.py.TEMPLATE:  Changed configuration entry name.
 - mysql_stat_run:  Replaced code to process JSON data with call to \_process_json.
 - mysql_stat_run:  Add to email body if email is setup.