+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 
 ## [2.3.0] - 2020-07-31
-- Updated to use the mysql_libs v5.0.0 library.
+- Updated to use the mysql_libs v5.0.2 library.
 
 ### Fixed
 - config/mysql.cfg.TEMPLATE:  Point to correct socket file.